Remove `Test` class
It seems to just be a wrapper around a function.  It's cleaner and more
efficient to just call the function directly.
